[FIX] sign: update options of selection sign items
Issue:
======
When you add a new selection sign item in a template and add new options
or update them it will raise an error.

Steps to reproduce the error:
=============================
- Open any editable sign template
- Add a selection field
- Click on the selection field
- Add new option in the options field

Origin of the error:
====================
In saas-16.4 option_ids contains a `REPLACE_WITH` command which is an
array of length 3 of this format `[SET, _, ids]` so each time we set the
new options with the new ids.
In 17.0 option_ids contains a list of `LINK` command where each one have
an id to link , so we will have an array of commands each of this format
`[LINK, id, _]`  that's why the old code doesn't work anymore.

Solution:
=========
We get all the ids of the `LINK` commands

opw-3596542

closes odoo/enterprise#51010

Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>