[FIX] document_hr_recruitment: fix document uploading from applicant
Description of the issue/feature this PR addresses:
n the kanban view of all applicants within the recruitment module, an issue
arises when applicants attempt to upload documents, this functionality is
currently malfunctioning.

Current behavior before PR:
Document uploads by applicants are failing.

Desired behavior after PR is merged:
Applicants should be able to successfully upload documents.

Fix:
The PR addresses the problem by modifying the default_res_model in the context.
This adjustment is crucial as the current setting results in an incorrect model
name, preventing the display of uploaded documents in this model.

task-3579833

closes odoo/enterprise#57840

X-original-commit: 5f99d82ae39d50554ba04cef7dfc990d39716ae9
Related: odoo/odoo#155958
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>