[FIX] pos_blackbox_be: fix wrong override in product screen
Prior to this commit, the override made on the method _setValue in
the product screen was including a call to the parent method without
the correct arguments leading to no behaviour when clicking on the
numpad in the product screen when the pos_blackbox_be module was
installed (e.g. when trying to do a discount). This commit fixes this
error.

closes odoo/enterprise#51168

X-original-commit: 5c9be8c34013ec9e31eaa7e143c4cbd7eef397ef
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>