[FIX] account_reports: Fix report line label in report form view
Before this commit we use column.name as label of our columns
for report lines in the report form view.
It caused that the label was in lower case.
The aim of this commit is using the label instead of the name to
have the column label well formatted.

no task id

closes odoo/enterprise#51088

X-original-commit: 80687287b6f10befcc2f2bb170a6bf10b0b0c8c2
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>