[FIX] : iot: change local storage implementation
Previously, each report-device link was stored as a unique entry
in the local storage. Now, a structured object in the local
storage contains these links. Therefore, the need to iterate
through all local storage entries for display purposes has been
eliminated.

The entry in localStorage "odoo-iot-linked_reports" is the JSON of an
object where the key is the report id and the data is a list of the
ids of the printers linked to it

closes odoo/enterprise#51675

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>