[FIX] whatsapp: prevent location variable deletion on header type change
Steps to reproduce:
1. Install `WhatsApp`
2. Create a new template and set the header type to `location`
3. Change the body and try to save the template
4. Validation error will be raised

When the header type is changed from `location` to any other type then the
record for location variable IDs is not deleted. Also, when the header type
is set to the `location` and some changes are made in the body, records of the
`location` variable are automatically deleted.

`_compute_variable_ids` deletes variable IDs if the header type is set to
`location` and there are existing header location variables.

Solution:
Deleting existing header location variables only if the header type is changed
from `location` to any other type.

Task-3520420
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: dde34465ba0c359a9f5ae1054b2e20e0a32ea8e2
Part-of: odoo/enterprise#52905