[IMP] whatsapp: sequencing template variables
Before This Commit:
Currently, the variables are not populating in the proper sequence. It makes the
variable tab look untidy.

After This Commit:
Sequence the variables with variable type in the sequence
Location Defaults > Header > Buttons > Body Variables

Solution:
used the _order attribute to sort the variables in the order of variable type
and placeholder
 1. location - address
 2. location - latitude, longitude
 3. location - name
 4. header
 5. button
 6. body

Task - 3516187

closes odoo/enterprise#51270

X-original-commit: eeb8f60ac879e907cc372733b10c1ad498713b08
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>