[FIX] documents_account: fix creation of accounting document
Steps to reproduce:
- Install Accounting and Documents
- Go to Documents settings
- Activate "Accounting" option (Workspace: Finance) and add a journal
to synchronize (Journal: Customer Invoices - Workspace: Finance)
- Go to Accounting and create an invoice
- Click on Print button and select "Invoices"

=> The following Validation Error can be raised:
"The operation cannot be completed: This attachment is already a document"

(This issue depends on the speed of the platform and doesn't always happen)

Cause:
When printing the invoice, an attachment is created, but no document is created
in Documents immediately (as configured in Documents settings).
Once the attachment created, it is loaded in the chatter and several calls are
made to "register_as_main_attachment" method of the attachment.
This method is triggering a write on the related record (i.e. the invoice), which
tries to create a document (in Documents) from the attachment.
When trying to create the document, it first checks if a document already exists
for that attachment.
But in some cases, there can be 2 concurrent calls where the first one has not
finished creating the document when the second one checks if a document already
exists.
It then tries to create a second document for the attachment and triggers the SQL
unique constraint.

Solution:
The creation of the document should not depend on the loading of the attachment
in the chatter.
It should be created when the attachment is created.
So, when the first attachment of an account move is created, the related document
will also be created.

opw-3464593

closes odoo/enterprise#51530

X-original-commit: b9745a55eaa32104546185c4d19986e31dba9b95
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>