[FIX] knowledge: prevent layout overflow on editor
Currently, the right column of the Knowledge editor can overflow the
viewport when there is not enough space for the topbar. It can happen
when the user resizes the sidebar or types a long title in the input
text of the topbar.

Steps to reproduce the issue:
1. Open Knowledge on Desktop.
2. Select the title input field.
3. Type a long title.
=> The input field over expands and the right column overflows the window.

TO BE: The right column should never overflow the window. The input text
should shrink when there is no more space.

task-3507172

X-original-commit: ed2672f2388a2917f6f38fa6eed04a1a5747d7ba
Part-of: odoo/enterprise#51280