[FIX] website_sale_renting: correct the default dates
Steps to reproduce:
-------------------
- set a "Daily" period for a rented product;
- go to ecommerce on the product page;

Issue:
------
The default period is not correct.
The dates and duration don't match.

Cause:
------
For a product whose rental is counted in days,
the duration is the number of days selected.
For a product that can be rented by the hour,
The duration is calculated using the difference
between the datetimes.

Solution:
---------
When the product can be rented in days, take the whole day,
i.e. 00:00:00 on the first day to 23:59:59 on the last day.

After this commit:
If the default period is 1 day and it's the 21st of the month,
the dates will be from the 22nd to the 22nd with a duration of 1 day
(on the datepicker, only one day is selected).

opw-3573163

closes odoo/enterprise#51878

X-original-commit: 24e271749a6623cf39604570ce45aba89eb679b5
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>