[FIX] hr_contract_salary: need company to compute display_name
When you don't have the multi company group on the user that will create
an offer, you get a traceback when you try to create a new offer.

The issue is caused, because when it opens the form view of offer, it
will compute the display name and for this computation it needs the
currency_id which is a related to the company. As the company field is
not present in the view when you don't have the multi-company group,
the state sent by the onchange doesn't contains the company so it cannot
get the currency for the display name.

So we put this field in invisible for users that don't have the group to
be able to compute the display_name.

closes odoo/enterprise#51259

X-original-commit: d57793a87c583a092095eae9ac1fb16f9c71b162
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>