[IMP] knowledge: cache command palette permission check
When the user enters specific terms in the command palette, and there is
no article matching those terms, the system will display a helper that
allows the user to quickly create a new article with the provided name.
The helper will only be shown if the user has the permission to create a
new article in the database. To verify this permission, the system will
perform an RPC call to the server each time the helper needs to be displayed.

In practice, it's highly unlikely for the user to lose the 'create'
privilege between two searches. Therefore, one can cache the result of
the RPC call that checks if the user is allowed to create a new article.
Caching this result should reduce the number of RPC calls triggered when
the user types something in the search bar.

To implement the caching mechanism, we will simply use the `memoize`
utility function to cache the promise checking the user's 'create'
privilege. The system will perform the RPC call on the first
invocation, cache the promise, and reuse it for subsequent calls,
thereby improving speed.

task-3603346

closes odoo/enterprise#53740

X-original-commit: 0143d054449d0e207d423487650e18a9279347ec
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>