[FIX] documents_spreadsheet: list test fails in some timezones
The test `list with a contextual domain` was failing on some timezone
because the date isn't well patched. We cannot correctly patch the date
at the moment, because PyDate doesn't use luxon but new Date().

Fix the issue by setting the date to 12PM, so the test will work in
most timezones. We should properly fix the issue once PyDate make
the switch to luxon.

closes odoo/enterprise#51306

X-original-commit: 80660cb604613c1f067f847fe46e7d03dbb5e762
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>