[IMP] account_followup: additional followers for partner chatter
Follow-up emails only appear in the partner chatter (since follow-ups are not their own dedicated model).
The same holds for the (potential) reply of the partner to the email.
If the accountant does not follow this chatter they may miss the reply.

This commit adds a new option to the follow-up levels that allow adding users as followers of the partner chatter
(when the follow-up level on which the option is set is executed).

task-3604569

closes odoo/enterprise#52101

Related: odoo/odoo#144945
Signed-off-by: William André (wan) <wan@odoo.com>