[FIX] voip: give valid domain format in search records in voip_call
This traceback arises when the user tries to call Recent.

To reproduce this issue:
1. Install 'VoIP'
2. Open Discuss/Softphone/Recent
3. Try to search for something in the Voip search bar

Error: ' ValueError: too many values to unpack (expected 3)'

When the user tries to search for something in the search bar an invalid domain
format is formed in 'get_recent_phone_call'method.
Which leads to the traceback from here

https://github.com/odoo/enterprise/blob/5c42708683905cfb5fe8c1005aabbbc9fb016fa2/voip/models/voip_call.py#L79-L85

After applying this commit will resolve the issue by giving a valid
domain format to search.

sentry-4644858097

closes odoo/enterprise#51358

X-original-commit: 981c224178b4331da4ff7b5db7a1e8bfb2d54eae
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>