[FIX] pos_preparation_display: Fix pdis order going back to first stage
Prior to this commit, when two preparation display received the same
order, it was then impossible to delete this order as the order would
come back to the first stage when trying to delete it. This is now
solved by virtually hiding the order when it is deleted and really
delete it when the order is deleted in all preparation display.

We also change the invitation text to load demo data in the preparation
display.

closes odoo/enterprise#51555

X-original-commit: b4264b7e1accf4da1fb5e6c568732442d9cbf043
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>