[FIX] web_enterprise: fix topbar dropdown darkmode focus
This commit solves an issue about the dropdown toggle from the topbar
navigation not being highlighted on `:focus` state.

This comes from a color issue, the `background-color` of the toggle
being the same as the navbar one in dark mode.

To solve this issue, we simply override the variable in dark mode, and set it
to `$o-gray-300`.

task-3607749

closes odoo/enterprise#54045

X-original-commit: 46a21926a3b1c048929e8dca283313c10dbb4b60
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>