[FIX] whatsapp{,_account}: consider different account_move types
Summary
-------
If you create a Whatsapp template for the Journal Entry model, a error
can appear if you try to submit it for approval.

Steps to reproduce
------------------
* install `whatsapp_account`
* create a whatsapp template with the following values:
  - Applies to: Journal Entry
  - Header Type: Document
  - Report: Invoices
* try to submit it for approval

You should be met with the following message:
"Only invoices could be printed."

Cause
-----
The Journal Entry model (aka `account.move`) can be divided in two
types: generic Journal Entries and Invoices (bill, receipt, ...).
One difference between generic Journal Entries and Invoices is that
Journal Entries are not intended to be printed.

When a user submits a whatsapp template with the Header Type set to
document, a sample record is selected to generate a demo PDF. In our
case, that record will be an `account.move`. The issue occurs because
that sample record can be a generic Journal Entry (i.e. not printable).

Fix
---
This commit ensures that the selected sample `account.move` is always an
invoice.

opw-3543875

closes odoo/enterprise#51841

X-original-commit: 8ec066f9a18cdffe3cb32805e31f33e09d44d9d6
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>