[FIX] mrp_workorder: shop floor suggest worksheet improvement empty WO
To reproduce:
- Create a MO with a WO with no steps
- Open the WO card in shop floor and click suggest worksheet improvement
in the card's menu
- Click any of the options

Current behavior: An empty dialog box appears.

Expected behavior: For the add a step button, it should continue to the
wizard to create the step, this will be forward ported in another fix.
The other buttons should not appear if there are no steps for the
selected WO.

closes odoo/enterprise#50148

closes odoo/enterprise#52456

Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Claeys Pieter (clpi) <clpi@odoo.com>