[FIX] industry_fsm_sale: enable quotation creation from task
Issue:
In a mono-company setup, users trying to create a quotation from a task
face an error message because the mandatory 'company' field in the
'other info' tab is not visible, unlike in multi-company setups.

Steps to Reproduce:
1. Install the Field Service app.
2. Activate 'Extra Quotations' in the Field Service general settings.
3. Enable 'Extra Quotation' in the settings of the relevant project.
4. Try creating a quotation from a task in a mono-company environment.
5. Encounter an error.

Resolution:
Modified the functionality so that the 'company_id' field defaults to
the current company.

opw-3596984

closes odoo/enterprise#52057

X-original-commit: 223b24572146bd799f48a8069bb5a54d43bd2d9c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Kawtar Drissi El Bouzaidi (kdeb) <kdeb@odoo.com>