[FIX] pos_settle_due: resolve partner creation error
Prior to this commit, inside the Point of Sale, clicking the Create
button on the partner screen to create a new partner resulted in an
error, preventing the creation process. This issue was traced back to
the recent refactor of the 'formatFloat' function (https://github.com/odoo/odoo/commit/054ca0a19aaf297f420a1b478b93ae26f1b943b8).
The refactor changed '(value || 0)' to 'value', leading to problems
when the value is undefined.

This commit resolves the issue by ensuring that 'undefined' values are
properly handled and prevented from being sent during the partner
creation process.

opw-3594167

closes odoo/enterprise#51332

X-original-commit: b29a2763fd2a8a812a5f038bc63ebc7971736dee
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>