[FIX] hr_appraisal: make use of the right ribbon color
This commit replaces the use of `secondary` for the canceled state of an
appraisal for a `danger` background.

Prior to this commit, there was a readability issue about a ribbon having
a `bg-secondary` rather than a `text-bg-secondary` one, inside the
`hr_appraisal` module. There was also an issue about the color used which
could be misleading, making the tag look like if the appraisal was in a
draft state rather than being canceled.

To fix these two issues, we apply a `text-bg-danger` class to the ribbon.

task-3594929

closes odoo/enterprise#51295

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>