[FIX] spreadsheet_edition: spreadsheet record name should be a string
When fetching the record data from the server via `join_spreadsheet_session`,
we rely on the field `display_name` while expecting it to be a string.
This will however not be the case since [1] (i.e. if `_rec_name=""` then
`display_name=False`). The client components are expecting a string, we
therefore fix `join_spreadsheet_session` accordingly.

[1] https://github.com/odoo/odoo-security/commit/c68fb4a494acc38b97678ed2a1fe18dbf8f3fa9d

closes odoo/enterprise#51479

Task: 3562167
X-original-commit: 41882c538ac7b62058750450408f4076c5a2529f
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>