[FIX] l10n_be_hr_payroll_fleet: convert fuel type to lambda
https://github.com/odoo/enterprise/pull/47717 fixed the display of the fuel
type for a new car, as it was not being shown. However, the change from that PR
was not respecting stable policy. This commit fixes this by changing the
selection values from a static list to a lambda function that returns that list
This makes the selection values not be stored in the database.

closes odoo/enterprise#51389

X-original-commit: 68b7f522b88ea82068087873f5fb5ff9dd8b17b9
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>