[FIX] delivery_sendcloud: allow to duplicate a carrier
send cloud shipping products, records created automatically and that should be
linked to only one carrier are linked to two carriers if we use the duplicate
button.

steps to reproduce:
- configure a sendclound carrier (carrier A)
- duplicate it (carrier B)
- on B, click on "Load your SendCloud shipping products"

before this commit:
- On A, sendcloud shipping product is deleted. This is because of

https://github.com/odoo/enterprise/blob/2fcc1224915301b4cbd1c582a319d37961d1120f/delivery_sendcloud/wizard/sendcloud_shipping_wizard.py#L60

after this commit:
- sendcloud products are not copied when using copy(). Adding an api.constrains
on top of that to ensure that a sendcloud product is not used on another
carrier seems overkill.

opw-3573247

closes odoo/enterprise#52346

X-original-commit: 621076418a70b81bdfc15d855e925f2c82e3513c
Signed-off-by: William Henrotin (whe) <whe@odoo.com>