[IMP] core: better lang check
Before this commit, we check if language is valid and active when `env.lang`.
1. SQL f'->>{env.lang}' is safe
2. `env.lang` is frequently called and slow

This commit
a) removes language check for `env.lang` to improve performance.
The field value reading time for translated field is reduced by 50% when the
field value has been cached.

b) adds language check when necessary
currently when you read/write in inactive or invalid language, you will actually
read/write in en_US. It would be safer to raise an error instead of doing
everything in en_US especially when some rpc read/write script forgets a
language is inactivated in the server

For feature, the added language validation happens when
1. `write` translated fields
2. `create` records with translated fields
3. `fetch` field values for translated fields
For performance the added language validation shouldn't happen when
1. `read` cached translated field value

c) for sake of a) the `env.lang` is not safe anymore
the following commits refactor SQL queries to avoid SQL injection with the help
of `_field_to_sql`

Part-of: odoo/enterprise#51115