[FIX] l10n_be_hr_contract_salary: default car in wizard
When you open the wizard to select generate a link, the car_id is auto
completed throught the compute field.

The compute takes the first fleet.vehicle where the employee or
applicant is the driver or future driver, but does not check that the
vehicle is of type car.

So when the employee has a car and a bike or just a bike, it can put the
bike in the car_id field (event if the domain on the field prevent from
selecting anything else than car)

closes odoo/enterprise#51373

X-original-commit: 5c7681868de66c500a5ab7edaa732749018acb87
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>