[IMP] stock_barcode: scan dest./package behavior
Before this commit, in the Operations Types Barcode Apps configuration,
if the setting for Destination Location or Put in Pack was set on "After
group of Products", once in the app, if the user scans multiple products
then scan a destination, the destination is actually applied only to the
last scanned product.
For the put in pack, all the scanned product were placed in the package,
except when an existing package is scanned (in this case, like for the
destination, only the last scanned product was concerned.)

This worked like that because in fact, this setting's option reflected
the old optional setting and should work like before the creation of
those settings.
But the label, "After group of Product", was misleading and it could be
verify interesting to make those settings work like the label said and
applied the destination and the package after a group of scanned
product.

This commit do that. But this change implies an issue: what should
happen if you scan partially a reserved line then scan a destination or
a package ? Should all the quantity be moved there ?
To avoid this issue, we split the reservation before to assign the
package or destination on an uncompleted reserved line.

task-3599397

Part-of: odoo/enterprise#51603