[FIX] l10n_es_reports: Libros export unhandled group tax
When looping over the tax in report line's tax_ids, sometimes the tax we
receive is a group tax, which consist of 2 children tax inside it. They
all have different id, and that causes the sheet_line_vals dictionary
that depend on the id to be inaccurate.

This commit register all of the children inside the tax group to the
sheet_line_vals, by applying a `flatten_taxes_hierarchy` on the tax_ids

This commit also adds an improvement where tax with l10n_es_type of
'ignore' are now ignored (for example: DUA's 'Subtract Base' line),
and DUA line should have a type of 'F5' instead in `exp_line_vals`

closes odoo/enterprise#51366

Task-id: 3604971
Related: odoo/odoo#143314
Signed-off-by: Josse Colpaert <jco@odoo.com>