[FIX] mps: wrong filter in mps when coming from BOM
When clicking on "schedules" from the BoM view, the domain in the
search bar is correctly set, but the product/components displayed
are not filtered.

The SearchModel used in this fix https://github.com/odoo/enterprise/pull/51365 is no
longer possible as it has been removed in https://github.com/odoo/enterprise/commit/9b95c443dad042c1db8266c22fb6cf74a5ee179c

The solution implemented here is to pass a domain when launching the client action.

opw-3548470

closes odoo/enterprise#51528

X-original-commit: 4b430053d5d84939bc99030b09681930a5912699
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>