[FIX] account_accountant: apply analytic model in bank recon
The analytic distribution model applies to journal entries but not to entries created in the bank recon widget. This fix applies the distribution model based on the partner, partner category and account prefix.

To reproduce:
- create a distribution model that applies to a particular account prefix
- create an invoice with a line that uses an account containing the prefix in the code (the distribution model applies)
- In bank reco, on the manual operations tab, use the same account (the distribution model doesn't apply)

Solution:
Override the `_compute_analytic_distribution` method from analytic.mixin in `bank.rec.widget.line`
In order for this to work, the corresponding community PR removes the `precompute` attribute.

manual forward port of #51343

OPW-3558060

closes odoo/enterprise#53025

X-original-commit: 666efcc36f6e8efb6ce2c4f6cbd2c7bafb6d64ad
Related: odoo/odoo#146757
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>