[FIX] industry_fsm_sale: fix "time & material" display in portal tasks
In field service's portal task view, a display error would occur with
the "Time & Material" section ; it would get displayed next to the
"Timesheets" section which is wrong, as it should be displayed on top of
it.

The issue is that with the rework of portal's front in 17.0, the xpath
expression that supposedly selects the "Timesheets" section now selects
the "Timesheets" navigation menu, and makes the "Time & Material"
section display right on top of it. To fix this, the xpath expression
was changed to select the right element.

task-3609165

closes odoo/enterprise#51374

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>