[IMP] industry_fsm_stock: improve product catalog for tracked product
This commit's purpose is to improve the visual when a user adds a
tracked product to a fsm task. Currently, when the user clicks on the
'add' button to add the first product, or when he clicks on the
+/-/input section to modify the quantity of already added product, the
quantity is modified according the changes of the user, before opening
the wizard. This can be misleading, since those changes are irrelevant,
and only the changes done through the wizard will impact the products of
the task.
To solve this issue, the change made by the button are no longer
visible, and the wizard is opened directly. Besides, the 'input' section
has been change to be a button instead in order to open the wizard
instead of letting the user writting useless stuffs in the input
section.

task-3600870
https://www.odoo.com/web#id=3600870&menu_id=4720&cids=1&action=333&active_id=4105&model=project.task&view_type=form

closes odoo/enterprise#51447

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>