[FIX] sign: sign document in mobile
Issue:
======
After adding all the values in a sign doucent and then go back to
change other fields, clicking in a text field and then clicking on
another will make the sign & validate button unresponsive.

Steps to reproduce the issue:
=============================
- install sign
- send a sign request to a customer , the sign document should have at
  least 2 text fields
-
- Open the sent link  using mobile display and add all tha values but do not validate
- Click on any text field
- Click on another text field and click next
- Clicking on the button validate will do nothing.

Origin of the issue:
====================
Toggling between text fields will remove `show` class from the old
bottom sheet and add it to the new one. but when closing the new one ,
the old one will still exist but we can't see it  because it doesn't
have the class `show` but it's display is still `block` , that's why
clicking on the button validate woudln't work , it's below the
div of the bottom sheet.

Solution:
=========
Removing the manual style as block and adding it to the show class so it
will be updated automatically when we add/remove the class.

opw-3477410
opw-3511365m

closes odoo/enterprise#51581

X-original-commit: e33dc15c5dc87e05b6a6fedb7f7c695d73132aa4
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>