[FIX] l10n_de_reports: fix document field in datev export
Belegfeld (attachment field) is usually the same as the move name.
However, right now if a user enters a custom Bill reference on a Vendor Bill,
the Belegfeld will contain the reference instead.
All attachments are generated using the move name,
so the connection between the the datev report line and the attachment is lost during export.
This commit changes this behaviour by keeping the move name
as the Belegfeld when exporting the csv file.

opw-3589455

closes odoo/enterprise#51770

X-original-commit: 4128d60507abb7f82949c069d35a3dcd3f774d19
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>