[FIX] account_followup: use the `email_from` to send followup emails
Before this fix, the `email_from` set on the `mail.template` is not used
when sending a followup email. Instead, only
`partner._get_followup_responsible()` is used to set the author and the
sender of the email (see the `_send_email` function).

Now, we read the `email_from` to set the sender of the email. We use
`_render_template` to possibly evaluate a dynamic expression (for
instance: `{{ object._get_followup_responsible().email_formatted }}`).

opw-3574286

closes odoo/enterprise#52488

X-original-commit: 40c4f21f63f8e2829f03a1ce39c2bd54e16828f5
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>