[FIX] industry_fsm_stock: display the on hand stock for right warehouse
Steps to reproduce:
- Install `industry_fsm_stock`.
- Create a new warehouse.
- Assign you a default warehouse on your user's profile.
- For a defined product ensure there is stock in the 2 warehouses.
- Create a new FSM task, and click on the add product stat button.
- You should see that the On Hand stock displayed if the stock of all
  warehouse.

Cause:
This issue happen because, when retreiving the `qty_available` for the
products, the context doesn't contain the warehouse that should be
used, so the method `_compute_quantities` get quantities from all
warehouse.

Solution:
Override `action_fsm_view_material` in industry_fsm_stock to add the
user default warehouse in the context.

opw-3516524

closes odoo/enterprise#52444

X-original-commit: 23dd614761160d187499d95922e6be048603656e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>