[FIX] web_mobile: crash on back button in mobile apps
During the removal of jQuery BBQ library [1], an `Event.getState()` call
was unfortunately left.

This resulted in a crash when using the device's back button/gesture,
which triggers a 'hashchange' event, but only in the mobile apps (iOS
and Android).

This commit properly converts the related event handling to use the
native window.location API.

Steps to reproduce:
- Open the backend using the mobile app
- Go to an app
- press the back button/execute the back gesture
=> error: e.getState is not a function

opw-3602245

[1]: odoo/odoo@507c36883675a45a15485ce5c73aa4557ba23639

closes odoo/enterprise#51722

X-original-commit: f69c1e219a2163988416fd8e45f8a50308f3e2f7
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>