[FIX] sign: drag&drop fields doesn't work on mobile devices
In the Sign app, the user can drag & drop fields onto a signature
template. Sadly this feature isn't properly supported on mobile devices
with touch support (like tablets) but nothing prevent the user from
attempting to use it, leading to errors being thrown.

Supporting this feature on those device being out of scope for a fix in
a stable release, this commit disallow to edit on mobile OS (i.e.
iOS/Android) to avoid user confusion.

Steps to reproduce:
- To to sign app on a mobile device (like a tablet)
- Open a editable template
- Try to drag&drop a field from the sidebar to the document viewer
=> error is thrown or nothing happens

opw-3582695

closes odoo/enterprise#52941

X-original-commit: 1e37c06b5ea8d596c8b86ff6e7b9a007435ac3b5
Signed-off-by: Pierre Paridans (app) <app@odoo.com>