[FIX] whatsapp: update template count on sync templates
When the user tries to sync all the templates from `WhatsApp Business Account`
the number of templates in the stat button is not updated.

Steps to produce:

- Create a `WhatsApp Business Account`
- Click on `Sync Templates`

Problem:

The compute function `_compute_templates_count` is responsible for updating the
`templates_count` field. But this compute function has no dependencies on it. So
it will be called when the page is reloaded.

Solution:

The function `button_sync_whatsapp_account_templates` returns a
`display_notification` for displaying the Template Created and Updated count.
After this, we will close the `display_notification` using
`ir.actions.act_window_close` using `next`.

Task - 3619128

closes odoo/enterprise#56395

X-original-commit: 1a01fc737fed8cff056aec537d26ad4047ae0744
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>