[FIX] website_helpdesk: give access token to customer attachment
Issue:
------
When we create a ticket from website form with an attachment,
the portal user no longer has access to his attachment.

This is problematic because he may think that it has not been sent correctly.

Cause:
------
The document has no `access_token`.

Solution:
---------
Give an `acces_token` to the document linked to the ticket.
In this way, the document will appear in the portal view's chat.

opw-3601205

closes odoo/enterprise#51702

X-original-commit: aacf90852cfcba71ba4b62ceaefecf7da723ebf3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>