[FIX] web_enterprise: solve undeterministic list isStudioEditable
Before this commit, the clickAll tour crashed (https://runbot.odoo.com/runbot/build/54182109 (version 17.0 and after, but the fix is relevant before those versions)

This was -- probably -- due to:
- a list renderer is opened, the web_enterprise upsell studio feature listens on the actionService to determine if the current action can be editable with studio.
- a client action is mounted
- just before the list renderer stops listening (which happens at unmount time) the same function of the first list renderer triggers.
- at this point the action that is evaluated is the one of the client action, which crashes because res_model is not defined on client actions definitions.

The fix is relevant in 16.0 because it solves a case were a second act_window action, in target new, would spawn with a list renderer too, making the first one trigger its function without crash but in the wrong conditions.

Also, the studio upsell feature has been disabled when an action is in target new (in a dialog). This is fine functionally as a dialog is a secondary flow.

closes odoo/enterprise#52173

X-original-commit: a3ce5c8864486b72662028847ba90d17d7ed427d
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>