[FIX] account_bank_statement_import_csv: import_fields 'amount' not in list while importing CSV file.
This issue is coming from  importing a CSV File in the
'account.bank.statement.line' model was caught by a sentry. Because
of if CSV file is not two columns,which are 'debit' and 'credit' , that time
getting error is 'amount' is not in the list.

step to reproduce-
- Install the Accounting.
- Go to Dashboard > click on import statement button of Bank card
- Upload the statement
- After, click on Test button
- Error will be generated.

Sentry-3931578971

closes odoo/enterprise#73525

X-original-commit: 02f23792c5e9d5e6258ff5edb5289e4b929806b4
Signed-off-by: William André (wan) <wan@odoo.com>