[FIX] web_studio: new reports have an oe_structure in the page
On the report's kanban inside studio, make a new report.
Add a a structure via the OdooEditor's Powerbox, for example 2 columns.

Before this commit, only the first columns was saved.
This was because of (but largely not entirely[1]) the created report only had
a <page /> node inside which one could add content. It was problematic
because the <page /> is the root element of the report document, and behaves as a block
*inside* which it is possible to add stuff (it is branded).
When adding a structure like the 2 Columns, the odoo editor wraps the node of the current
selection and puts it inside the new structure.
In those cases, the branded block is put inside the first column.
When saving, the columns have disappeared.

After this commit, new reports have a div inside the root branded node, which
allows for changing it.

[1]: The real issue comes from the fact that branded nodes are considered as solid blocks
editable only on their inside, not around them. Even a tag change won't work.
This is a fundamental problem which we are still discussing

X-original-commit: 74e0e56df06f66a15bffff21d338fcc84a4b7b3b
Part-of: odoo/enterprise#53101