[FIX] web_studio: report editor: prevent inlining recursive t-call
Have a report that calls a template, which calls itself (all that via a t-call).

Before this commit, this case was not handled, leading to crashes (hard recursion limit errors)

After this commit, this case is protected against, the recursion is stopped before the second time
the template is called in the same stack.

opw-3505383

Part-of: odoo/enterprise#52884