[FIX] web_studio: report editor: prevent inlining recursive t-call
Have a report that calls a template, which calls itself (all that via a t-call).

Before this commit, this case was not handled, leading to crashes (hard recursion limit errors)

After this commit, this case is protected against, the recursion is stopped before the second time
the template is called in the same stack.

opw-3505383

X-original-commit: 82bb63d444bc15267681c6c35f03e42694e655f3
Part-of: odoo/enterprise#53101