[IMP] l10n_co_edi: setup l10n_co_edi_ref_nominal_tax for IBUA taxes
In 15.0 we introduced IBUA and ICUI taxes for non-healthy products.
We used the Volume field (regardless of the UoM) because we had
no choice, in stable we can't add fields.

In master we're adding the l10n_co_edi_ref_nominal_tax field.
Together with the fixed amount of the tax, we can use it to
compute back the sugar rate of the product.
It has also been added to the product view.

Fixed the missing tax groups. The fix will be backported saas-16.4.
Chart template did manually insert the groups, but it has to be done
through CSV files that are easier to be updated and checked
by non-technical people.

Previous task: https://www.odoo.com/web#id=3569752&model=project.task$
Task link: https://www.odoo.com/web#id=3614333&&model=project.task
task-3614333

closes odoo/enterprise#51492

Signed-off-by: Josse Colpaert <jco@odoo.com>