[IMP] web_studio: add assertions in a test.
The test "edit/delete menus" sometimes failed on runbot (error
23054) because there're 3 .btn_primary in the footer, whereas only
one is expected. When this happened, it was because the second
dialog wasn't closed yet. This commit adds assertions on the number
of dialogs, which helped to understand and reproduce the issue, and
which can also help in the future.

Related to PR odoo/odoo#143813 which fixes the issue.

closes odoo/enterprise#51804

X-original-commit: 9061dded31383be13a144d53e020ea53407ae437
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>