[IMP] web_gantt, planning: hook allocate hours
In this commit we introduce a hook in the computation of the gantt pill
title. Allocated hours display are computed in web_gantt, whereas it
is only introduced in other modules. A hook is then created here in
order to control the display of allocated hours in other modules. We use
this hook in this commit in planning in order to prevent the display
of allocated hours in the pills title when they are equal to 100% of
the available time.

taskid:3186636

closes odoo/enterprise#41943

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>